+++ casbin/persist/__init__.py
@@ -12,7 +12,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .adapter import *
-from .adapter_filtered import *
-from .batch_adapter import *
 from .adapters import *
