+++ casbin/persist/adapters/batch_adapter.py
