+++ tests/__init__.py
@@ -12,15 +12,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from . import benchmarks
-from . import config
-from . import model
-from . import rbac
-from . import util
+from . import benchmarks, config, model, rbac, util
 from .test_distributed_api import TestDistributedApi
 from .test_enforcer import *
 from .test_fast_enforcer import TestFastEnforcer
-from .test_filter import TestFilteredAdapter
+from .test_filter import TestFilteredFileAdapter
 from .test_frontend import TestFrontend
 from .test_management_api import TestManagementApi, TestManagementApiSynced
 from .test_rbac_api import TestRbacApi, TestRbacApiSynced
